PP-8831 Send phone number to stripe with international code
Stripe rejects phone numbers without an international code. Format the
phone number entered for the responsible person to include the UK
dialling code if no international dialling code is provided.
