--- app/utils/validation/telephone-number-validation.test.js
@@ -1,37 +1,0 @@
-'use strict'
-
-const { expect } = require('chai')
-
-const { invalidTelephoneNumber } = require('./telephone-number-validation')
-
-describe('telephone number validation', () => {
-  it('should return false for telephone number with spaces', () => {
-    const validPhoneNumber = '0113 496 0000'
-    expect(invalidTelephoneNumber(validPhoneNumber)).to.be.false // eslint-disable-line
-  })
-
-  it('should return false for telephone number with dashes', () => {
-    const validPhoneNumber = '0113-496-0000'
-    expect(invalidTelephoneNumber(validPhoneNumber)).to.be.false // eslint-disable-line
-  })
-
-  it('should return false for telephone number with mixed format', () => {
-    const validPhoneNumber = '(0113) 496 / 0000'
-    expect(invalidTelephoneNumber(validPhoneNumber)).to.be.false // eslint-disable-line
-  })
-
-  it('should return false for telephone number with international mixed format', () => {
-    const validPhoneNumber = '+36 / (1) 496 - 0000'
-    expect(invalidTelephoneNumber(validPhoneNumber)).to.be.false // eslint-disable-line
-  })
-
-  it('should return true for invalid telephone number', () => {
-    const validPhoneNumber = 'abc'
-    expect(invalidTelephoneNumber(validPhoneNumber)).to.be.true // eslint-disable-line
-  })
-
-  it('should return true for telephone number is less than 9 digits', () => {
-    const validPhoneNumber = '12345678'
-    expect(invalidTelephoneNumber(validPhoneNumber)).to.be.true // eslint-disable-line
-  })
-})
